Removing unused property causing dev-requirement for cyclical dependency
That `$_element` property (which was marked as Zend_Form_Element type,
which is why I had added the dev dependency on Zend_Form) isn’t
referenced in the FormErrors helper, or any of its parents. I also
looked through all ZF1 code for reference to that helper (to see if
anything else extends it), but couldn’t find anything.

Also changing the psr-0 namespace for this lib to just Zend_View_ (for
performance in non-optimized situations).  Had to add `Zend_View` as a
classmap for this though since it doesn’t have the trailing underscore
after “View”.

Figured I’d test code climate here as well, so added its coverage
reporter to the travis build.